refactor(test): split client.spec.ts (753 lines) by concern
DOM ハーネスを client.test-helpers.ts へ抽出し、機能ごとに 4 ファイルへ分割:
- client.mount.spec.ts (iframe マウント・URL パラメータ)
- client.requests.spec.ts (リクエスト／レスポンス・タイムアウト・可視性)
- client.crypto.spec.ts (nip44 / nip04 namespaces)
- client.validation.spec.ts (result 型検証・ライフサイクル・オプション検証)

テスト件数は不変。
