+++ packages/nosskey-iframe/src/client.requests.spec.ts
@@ -1,0 +1,268 @@
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { NosskeyIframeClient, NosskeyIframeError } from './client.js';
+import { createHarness } from './client.test-helpers.js';
+import type { Harness } from './client.test-helpers.js';
+import type { NostrEvent } from './types.js';
+
+describe('NosskeyIframeClient', () => {
+  let harness: Harness;
+
+  beforeEach(() => {
+    vi.useFakeTimers();
+    harness = createHarness();
+  });
+
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it('ready() resolves when a nosskey:ready message arrives from the iframe', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+
+    const readyPromise = client.ready();
+    harness.dispatchAsIframe({ type: 'nosskey:ready' });
+    await expect(readyPromise).resolves.toBeUndefined();
+    client.destroy();
+  });
+
+  it('getPublicKey() posts a request and resolves with the response result', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+
+    const pending = client.getPublicKey();
+    expect(iframe.contentWindow.postMessage).toHaveBeenCalledOnce();
+    const [request, targetOrigin] = iframe.contentWindow.postMessage.mock.calls[0];
+    expect(targetOrigin).toBe('https://nosskey.example');
+    expect(request).toMatchObject({
+      type: 'nosskey:request',
+      method: 'getPublicKey',
+      id: 'test-id-1',
+    });
+
+    harness.dispatchAsIframe({ type: 'nosskey:response', id: 'test-id-1', result: 'deadbeef' });
+    await expect(pending).resolves.toBe('deadbeef');
+    client.destroy();
+  });
+
+  it('getRelays() posts a request and resolves with the relay map', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+    const relays = {
+      'wss://relay.example': { read: true, write: true },
+      'wss://relay.read.example': { read: true, write: false },
+    };
+
+    const pending = client.getRelays();
+    expect(iframe.contentWindow.postMessage).toHaveBeenCalledOnce();
+    const [request, targetOrigin] = iframe.contentWindow.postMessage.mock.calls[0];
+    expect(targetOrigin).toBe('https://nosskey.example');
+    expect(request).toMatchObject({
+      type: 'nosskey:request',
+      method: 'getRelays',
+    });
+
+    harness.dispatchAsIframe({
+      type: 'nosskey:response',
+      id: (request as { id: string }).id,
+      result: relays,
+    });
+    await expect(pending).resolves.toEqual(relays);
+    client.destroy();
+  });
+
+  it('signEvent() posts the event and resolves with the signed event', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+    const unsigned: NostrEvent = { kind: 1, content: 'hi' };
+    const signed: NostrEvent = { ...unsigned, id: 'abc', sig: 'def' };
+
+    const pending = client.signEvent(unsigned);
+    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
+    expect(request).toMatchObject({
+      type: 'nosskey:request',
+      method: 'signEvent',
+      params: { event: unsigned },
+    });
+
+    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: signed });
+    await expect(pending).resolves.toEqual(signed);
+    client.destroy();
+  });
+
+  it('rejects with NosskeyIframeError when the response carries an error', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const pending = client.getPublicKey();
+    const iframe = harness.iframes[0];
+    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
+
+    harness.dispatchAsIframe({
+      type: 'nosskey:response',
+      id: request.id,
+      error: { code: 'USER_REJECTED', message: 'nope' },
+    });
+
+    await expect(pending).rejects.toBeInstanceOf(NosskeyIframeError);
+    await expect(pending).rejects.toMatchObject({ code: 'USER_REJECTED', message: 'nope' });
+    client.destroy();
+  });
+
+  it('rejects after timeout when no response arrives', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+      timeout: 50,
+    });
+    const pending = client.getPublicKey();
+    // Attach a catch handler immediately so the rejection is not treated as unhandled
+    // before the test assertion runs.
+    const settled = expect(pending).rejects.toThrow(/timed out/);
+    await vi.advanceTimersByTimeAsync(60);
+    await settled;
+    client.destroy();
+  });
+
+  it('ignores messages whose source is not the iframe', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const pending = client.getPublicKey();
+    const iframe = harness.iframes[0];
+    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
+
+    // Unrelated window sends a matching response — must be ignored.
+    harness.dispatch(
+      { type: 'nosskey:response', id: request.id, result: 'fake' },
+      'https://nosskey.example',
+      { unrelated: true }
+    );
+
+    // Subsequently, the real iframe replies and the promise resolves.
+    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'real' });
+    await expect(pending).resolves.toBe('real');
+    client.destroy();
+  });
+
+  it('ignores messages whose origin does not match the iframe origin', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const pending = client.getPublicKey();
+    const iframe = harness.iframes[0];
+    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
+
+    harness.dispatch(
+      { type: 'nosskey:response', id: request.id, result: 'spoof' },
+      'https://evil.example',
+      iframe.contentWindow
+    );
+
+    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'legit' });
+    await expect(pending).resolves.toBe('legit');
+    client.destroy();
+  });
+
+  it('destroy() removes the iframe, detaches the listener, and rejects pending requests', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+    const pending = client.getPublicKey();
+    const settled = expect(pending).rejects.toThrow(/destroyed/);
+
+    client.destroy();
+    await settled;
+
+    expect(harness.container.children).not.toContain(iframe);
+
+    // Post-destroy: further messages must not resolve anything.
+    harness.dispatchAsIframe({ type: 'nosskey:response', id: 'test-id-1', result: 'late' });
+    // No assertion needed — absence of unhandled rejections is the check.
+  });
+
+  it('toggles iframe visibility when a nosskey:visibility message arrives', () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+    expect(iframe.style.display).toBe('none');
+    expect(iframe.getAttribute('aria-hidden')).toBe('true');
+
+    harness.dispatchAsIframe({ type: 'nosskey:visibility', visible: true });
+    expect(iframe.style.display).toBe('block');
+    expect(iframe.getAttribute('aria-hidden')).toBeNull();
+
+    harness.dispatchAsIframe({ type: 'nosskey:visibility', visible: false });
+    expect(iframe.style.display).toBe('none');
+    expect(iframe.getAttribute('aria-hidden')).toBe('true');
+
+    client.destroy();
+  });
+
+  it('ignores nosskey:visibility messages from unrelated sources', () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    const iframe = harness.iframes[0];
+
+    harness.dispatch({ type: 'nosskey:visibility', visible: true }, 'https://nosskey.example', {
+      unrelated: true,
+    });
+    expect(iframe.style.display).toBe('none');
+    client.destroy();
+  });
+
+  it('request rejects when the iframe has no contentWindow', async () => {
+    const client = new NosskeyIframeClient({
+      iframeUrl: 'https://nosskey.example/iframe',
+      window: harness.window,
+      document: harness.document,
+      container: harness.container as unknown as HTMLElement,
+    });
+    // Simulate a not-yet-loaded iframe.
+    (harness.iframes[0] as unknown as { contentWindow: null }).contentWindow = null;
+    await expect(client.getPublicKey()).rejects.toThrow(/not ready/);
+    client.destroy();
+  });
+});
