--- packages/nosskey-iframe/src/client.spec.ts
@@ -1,753 +1,0 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { NosskeyIframeClient, NosskeyIframeError } from './client.js';
-import type { NostrEvent } from './types.js';
-
-/**
- * Minimal DOM harness. We construct a fake window/document/container so the
- * iframe's `contentWindow` is under our control and we can intercept
- * `postMessage` calls and drive the client's message listener directly.
- */
-interface FakeIframe {
-  src: string;
-  style: { display: string };
-  attributes: Record<string, string>;
-  contentWindow: { postMessage: ReturnType<typeof vi.fn> };
-  parentNode: FakeElement | null;
-  setAttribute(name: string, value: string): void;
-  getAttribute(name: string): string | null;
-  removeAttribute(name: string): void;
-}
-
-interface FakeElement {
-  children: FakeIframe[];
-  appendChild(node: FakeIframe): FakeIframe;
-  removeChild(node: FakeIframe): FakeIframe;
-}
-
-interface Harness {
-  window: Window;
-  document: Document;
-  container: FakeElement;
-  dispatch: (data: unknown, origin: string, source?: unknown) => void;
-  dispatchAsIframe: (data: unknown, origin?: string) => void;
-  iframes: FakeIframe[];
-}
-
-function createHarness(iframeOrigin = 'https://nosskey.example'): Harness {
-  const listeners: Array<(event: MessageEvent) => void> = [];
-  const iframes: FakeIframe[] = [];
-
-  const createElement = (tagName: string) => {
-    if (tagName !== 'iframe') {
-      throw new Error(`Unexpected createElement(${tagName})`);
-    }
-    const iframe: FakeIframe = {
-      src: '',
-      style: { display: '' },
-      attributes: {},
-      contentWindow: { postMessage: vi.fn() },
-      parentNode: null,
-      setAttribute(name, value) {
-        iframe.attributes[name] = value;
-      },
-      getAttribute(name) {
-        return iframe.attributes[name] ?? null;
-      },
-      removeAttribute(name) {
-        delete iframe.attributes[name];
-      },
-    };
-    iframes.push(iframe);
-    return iframe;
-  };
-
-  const container: FakeElement = {
-    children: [],
-    appendChild(node) {
-      container.children.push(node);
-      node.parentNode = container;
-      return node;
-    },
-    removeChild(node) {
-      const idx = container.children.indexOf(node);
-      if (idx >= 0) container.children.splice(idx, 1);
-      node.parentNode = null;
-      return node;
-    },
-  };
-
-  const doc = {
-    createElement,
-    get body() {
-      return container;
-    },
-  } as unknown as Document;
-
-  const win = {
-    addEventListener(type: string, handler: EventListenerOrEventListenerObject) {
-      if (type !== 'message') return;
-      listeners.push(handler as (event: MessageEvent) => void);
-    },
-    removeEventListener(type: string, handler: EventListenerOrEventListenerObject) {
-      if (type !== 'message') return;
-      const idx = listeners.indexOf(handler as (event: MessageEvent) => void);
-      if (idx >= 0) listeners.splice(idx, 1);
-    },
-    document: doc,
-    location: { href: 'https://parent.example/app/' },
-    crypto: {
-      randomUUID: (() => {
-        let n = 0;
-        return () => `test-id-${++n}`;
-      })(),
-    },
-  } as unknown as Window;
-
-  const dispatch = (data: unknown, origin: string, source?: unknown) => {
-    const event = { data, origin, source } as unknown as MessageEvent;
-    for (const handler of [...listeners]) {
-      handler(event);
-    }
-  };
-
-  const dispatchAsIframe = (data: unknown, origin = iframeOrigin) => {
-    const iframe = iframes[0];
-    if (!iframe) throw new Error('No iframe created yet.');
-    dispatch(data, origin, iframe.contentWindow);
-  };
-
-  return { window: win, document: doc, container, dispatch, dispatchAsIframe, iframes };
-}
-
-describe('NosskeyIframeClient', () => {
-  let harness: Harness;
-
-  beforeEach(() => {
-    vi.useFakeTimers();
-    harness = createHarness();
-  });
-
-  afterEach(() => {
-    vi.useRealTimers();
-  });
-
-  it('mounts an iframe with the WebAuthn allow attribute and hidden style', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    expect(harness.iframes).toHaveLength(1);
-    const iframe = harness.iframes[0];
-    expect(iframe.src).toBe('https://nosskey.example/iframe');
-    expect(iframe.getAttribute('allow')).toBe(
-      'publickey-credentials-get; publickey-credentials-create'
-    );
-    expect(iframe.style.display).toBe('none');
-    expect(iframe.parentNode).toBe(harness.container);
-    client.destroy();
-  });
-
-  it('appends embedded=1 and theme/lang query params when provided', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/#/iframe',
-      theme: 'dark',
-      lang: 'ja',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const iframe = harness.iframes[0];
-    const url = new URL(iframe.src);
-    expect(url.searchParams.get('embedded')).toBe('1');
-    expect(url.searchParams.get('theme')).toBe('dark');
-    expect(url.searchParams.get('lang')).toBe('ja');
-    expect(url.hash).toBe('#/iframe');
-    client.destroy();
-  });
-
-  it('omits theme/lang params when neither option is supplied', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/#/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const iframe = harness.iframes[0];
-    expect(iframe.src).toBe('https://nosskey.example/#/iframe');
-    client.destroy();
-  });
-
-  it('appends only theme when lang is omitted (and vice versa)', () => {
-    const themeOnly = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/#/iframe',
-      theme: 'auto',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const themeIframe = harness.iframes[0];
-    const themeUrl = new URL(themeIframe.src);
-    expect(themeUrl.searchParams.get('embedded')).toBe('1');
-    expect(themeUrl.searchParams.get('theme')).toBe('auto');
-    expect(themeUrl.searchParams.has('lang')).toBe(false);
-    themeOnly.destroy();
-
-    const langOnly = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/#/iframe',
-      lang: 'en',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const langIframe = harness.iframes[1];
-    const langUrl = new URL(langIframe.src);
-    expect(langUrl.searchParams.get('embedded')).toBe('1');
-    expect(langUrl.searchParams.get('lang')).toBe('en');
-    expect(langUrl.searchParams.has('theme')).toBe(false);
-    langOnly.destroy();
-  });
-
-  it('overrides existing embedded/theme/lang query params on the source URL', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe?embedded=1&theme=light&lang=en',
-      theme: 'dark',
-      lang: 'ja',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const iframe = harness.iframes[0];
-    const url = new URL(iframe.src);
-    expect(url.searchParams.getAll('theme')).toEqual(['dark']);
-    expect(url.searchParams.getAll('lang')).toEqual(['ja']);
-    expect(url.searchParams.getAll('embedded')).toEqual(['1']);
-    client.destroy();
-  });
-
-  it('resolves a relative iframeUrl against window.location.href', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: '/iframe',
-      theme: 'dark',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const iframe = harness.iframes[0];
-    // Harness location.href is 'https://parent.example/app/' so '/iframe'
-    // resolves under the same origin.
-    const url = new URL(iframe.src);
-    expect(url.origin).toBe('https://parent.example');
-    expect(url.pathname).toBe('/iframe');
-    expect(url.searchParams.get('embedded')).toBe('1');
-    expect(url.searchParams.get('theme')).toBe('dark');
-    client.destroy();
-  });
-
-  it("passes lang='auto' through to the iframe unchanged", () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/#/iframe',
-      lang: 'auto',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const iframe = harness.iframes[0];
-    const url = new URL(iframe.src);
-    expect(url.searchParams.get('embedded')).toBe('1');
-    expect(url.searchParams.get('lang')).toBe('auto');
-    client.destroy();
-  });
-
-  it('ready() resolves when a nosskey:ready message arrives from the iframe', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-
-    const readyPromise = client.ready();
-    harness.dispatchAsIframe({ type: 'nosskey:ready' });
-    await expect(readyPromise).resolves.toBeUndefined();
-    client.destroy();
-  });
-
-  it('getPublicKey() posts a request and resolves with the response result', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-
-    const pending = client.getPublicKey();
-    expect(iframe.contentWindow.postMessage).toHaveBeenCalledOnce();
-    const [request, targetOrigin] = iframe.contentWindow.postMessage.mock.calls[0];
-    expect(targetOrigin).toBe('https://nosskey.example');
-    expect(request).toMatchObject({
-      type: 'nosskey:request',
-      method: 'getPublicKey',
-      id: 'test-id-1',
-    });
-
-    harness.dispatchAsIframe({ type: 'nosskey:response', id: 'test-id-1', result: 'deadbeef' });
-    await expect(pending).resolves.toBe('deadbeef');
-    client.destroy();
-  });
-
-  it('getRelays() posts a request and resolves with the relay map', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-    const relays = {
-      'wss://relay.example': { read: true, write: true },
-      'wss://relay.read.example': { read: true, write: false },
-    };
-
-    const pending = client.getRelays();
-    expect(iframe.contentWindow.postMessage).toHaveBeenCalledOnce();
-    const [request, targetOrigin] = iframe.contentWindow.postMessage.mock.calls[0];
-    expect(targetOrigin).toBe('https://nosskey.example');
-    expect(request).toMatchObject({
-      type: 'nosskey:request',
-      method: 'getRelays',
-    });
-
-    harness.dispatchAsIframe({
-      type: 'nosskey:response',
-      id: (request as { id: string }).id,
-      result: relays,
-    });
-    await expect(pending).resolves.toEqual(relays);
-    client.destroy();
-  });
-
-  it('signEvent() posts the event and resolves with the signed event', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-    const unsigned: NostrEvent = { kind: 1, content: 'hi' };
-    const signed: NostrEvent = { ...unsigned, id: 'abc', sig: 'def' };
-
-    const pending = client.signEvent(unsigned);
-    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-    expect(request).toMatchObject({
-      type: 'nosskey:request',
-      method: 'signEvent',
-      params: { event: unsigned },
-    });
-
-    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: signed });
-    await expect(pending).resolves.toEqual(signed);
-    client.destroy();
-  });
-
-  it('rejects with NosskeyIframeError when the response carries an error', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const pending = client.getPublicKey();
-    const iframe = harness.iframes[0];
-    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-
-    harness.dispatchAsIframe({
-      type: 'nosskey:response',
-      id: request.id,
-      error: { code: 'USER_REJECTED', message: 'nope' },
-    });
-
-    await expect(pending).rejects.toBeInstanceOf(NosskeyIframeError);
-    await expect(pending).rejects.toMatchObject({ code: 'USER_REJECTED', message: 'nope' });
-    client.destroy();
-  });
-
-  it('rejects after timeout when no response arrives', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-      timeout: 50,
-    });
-    const pending = client.getPublicKey();
-    // Attach a catch handler immediately so the rejection is not treated as unhandled
-    // before the test assertion runs.
-    const settled = expect(pending).rejects.toThrow(/timed out/);
-    await vi.advanceTimersByTimeAsync(60);
-    await settled;
-    client.destroy();
-  });
-
-  it('ignores messages whose source is not the iframe', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const pending = client.getPublicKey();
-    const iframe = harness.iframes[0];
-    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-
-    // Unrelated window sends a matching response — must be ignored.
-    harness.dispatch(
-      { type: 'nosskey:response', id: request.id, result: 'fake' },
-      'https://nosskey.example',
-      { unrelated: true }
-    );
-
-    // Subsequently, the real iframe replies and the promise resolves.
-    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'real' });
-    await expect(pending).resolves.toBe('real');
-    client.destroy();
-  });
-
-  it('ignores messages whose origin does not match the iframe origin', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const pending = client.getPublicKey();
-    const iframe = harness.iframes[0];
-    const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-
-    harness.dispatch(
-      { type: 'nosskey:response', id: request.id, result: 'spoof' },
-      'https://evil.example',
-      iframe.contentWindow
-    );
-
-    harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'legit' });
-    await expect(pending).resolves.toBe('legit');
-    client.destroy();
-  });
-
-  it('destroy() removes the iframe, detaches the listener, and rejects pending requests', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-    const pending = client.getPublicKey();
-    const settled = expect(pending).rejects.toThrow(/destroyed/);
-
-    client.destroy();
-    await settled;
-
-    expect(harness.container.children).not.toContain(iframe);
-
-    // Post-destroy: further messages must not resolve anything.
-    harness.dispatchAsIframe({ type: 'nosskey:response', id: 'test-id-1', result: 'late' });
-    // No assertion needed — absence of unhandled rejections is the check.
-  });
-
-  it('toggles iframe visibility when a nosskey:visibility message arrives', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-    expect(iframe.style.display).toBe('none');
-    expect(iframe.getAttribute('aria-hidden')).toBe('true');
-
-    harness.dispatchAsIframe({ type: 'nosskey:visibility', visible: true });
-    expect(iframe.style.display).toBe('block');
-    expect(iframe.getAttribute('aria-hidden')).toBeNull();
-
-    harness.dispatchAsIframe({ type: 'nosskey:visibility', visible: false });
-    expect(iframe.style.display).toBe('none');
-    expect(iframe.getAttribute('aria-hidden')).toBe('true');
-
-    client.destroy();
-  });
-
-  it('ignores nosskey:visibility messages from unrelated sources', () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    const iframe = harness.iframes[0];
-
-    harness.dispatch({ type: 'nosskey:visibility', visible: true }, 'https://nosskey.example', {
-      unrelated: true,
-    });
-    expect(iframe.style.display).toBe('none');
-    client.destroy();
-  });
-
-  it('request rejects when the iframe has no contentWindow', async () => {
-    const client = new NosskeyIframeClient({
-      iframeUrl: 'https://nosskey.example/iframe',
-      window: harness.window,
-      document: harness.document,
-      container: harness.container as unknown as HTMLElement,
-    });
-    // Simulate a not-yet-loaded iframe.
-    (harness.iframes[0] as unknown as { contentWindow: null }).contentWindow = null;
-    await expect(client.getPublicKey()).rejects.toThrow(/not ready/);
-    client.destroy();
-  });
-
-  describe('nip44 / nip04 namespaces', () => {
-    const peerPub = 'b'.repeat(64);
-
-    it('nip44.encrypt sends the right request and resolves with ciphertext', async () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      const iframe = harness.iframes[0];
-      const pending = client.nip44.encrypt(peerPub, 'plain');
-      const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-      expect(request).toMatchObject({
-        type: 'nosskey:request',
-        method: 'nip44_encrypt',
-        params: { pubkey: peerPub, plaintext: 'plain' },
-      });
-      harness.dispatchAsIframe({
-        type: 'nosskey:response',
-        id: request.id,
-        result: 'cipher-payload',
-      });
-      await expect(pending).resolves.toBe('cipher-payload');
-      client.destroy();
-    });
-
-    it('nip44.decrypt sends ciphertext and resolves with plaintext', async () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      const iframe = harness.iframes[0];
-      const pending = client.nip44.decrypt(peerPub, 'cipher-payload');
-      const [request] = iframe.contentWindow.postMessage.mock.calls[0];
-      expect(request).toMatchObject({
-        method: 'nip44_decrypt',
-        params: { pubkey: peerPub, ciphertext: 'cipher-payload' },
-      });
-      harness.dispatchAsIframe({
-        type: 'nosskey:response',
-        id: request.id,
-        result: 'plain',
-      });
-      await expect(pending).resolves.toBe('plain');
-      client.destroy();
-    });
-
-    it('nip04 namespace mirrors nip44 shape', async () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      const iframe = harness.iframes[0];
-      const enc = client.nip04.encrypt(peerPub, 'p');
-      const [encReq] = iframe.contentWindow.postMessage.mock.calls[0];
-      expect(encReq.method).toBe('nip04_encrypt');
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: encReq.id, result: 'cbc?iv=z' });
-      await expect(enc).resolves.toBe('cbc?iv=z');
-
-      const dec = client.nip04.decrypt(peerPub, 'cbc?iv=z');
-      const [decReq] = iframe.contentWindow.postMessage.mock.calls[1];
-      expect(decReq.method).toBe('nip04_decrypt');
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: decReq.id, result: 'p' });
-      await expect(dec).resolves.toBe('p');
-
-      client.destroy();
-    });
-  });
-
-  describe('result type validation', () => {
-    const mkClient = () =>
-      new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-
-    it('getPublicKey rejects when the iframe returns a non-string result', async () => {
-      const client = mkClient();
-      const pending = client.getPublicKey();
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 42 });
-      await expect(pending).rejects.toThrow(/expected string result/);
-      client.destroy();
-    });
-
-    it('getRelays rejects when the iframe returns a non-object result', async () => {
-      const client = mkClient();
-      const pending = client.getRelays();
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'not-object' });
-      await expect(pending).rejects.toThrow(/expected object result/);
-      client.destroy();
-    });
-
-    it('signEvent rejects when the iframe returns a non-object result', async () => {
-      const client = mkClient();
-      const pending = client.signEvent({ kind: 1, content: 'hi' });
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 'not-object' });
-      await expect(pending).rejects.toThrow(/expected NostrEvent result/);
-      client.destroy();
-    });
-
-    it('nip44.encrypt rejects when the iframe returns a non-string result', async () => {
-      const client = mkClient();
-      const pending = client.nip44.encrypt('b'.repeat(64), 'plain');
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: {} });
-      await expect(pending).rejects.toThrow(/nip44\.encrypt: expected string result/);
-      client.destroy();
-    });
-
-    it('nip44.decrypt rejects when the iframe returns a non-string result', async () => {
-      const client = mkClient();
-      const pending = client.nip44.decrypt('b'.repeat(64), 'cipher');
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: 7 });
-      await expect(pending).rejects.toThrow(/nip44\.decrypt: expected string result/);
-      client.destroy();
-    });
-
-    it('nip04.encrypt rejects when the iframe returns a non-string result', async () => {
-      const client = mkClient();
-      const pending = client.nip04.encrypt('b'.repeat(64), 'plain');
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: null });
-      await expect(pending).rejects.toThrow(/nip04\.encrypt: expected string result/);
-      client.destroy();
-    });
-
-    it('nip04.decrypt rejects when the iframe returns a non-string result', async () => {
-      const client = mkClient();
-      const pending = client.nip04.decrypt('b'.repeat(64), 'cipher');
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      harness.dispatchAsIframe({ type: 'nosskey:response', id: request.id, result: false });
-      await expect(pending).rejects.toThrow(/nip04\.decrypt: expected string result/);
-      client.destroy();
-    });
-  });
-
-  describe('lifecycle and option validation', () => {
-    it('ready() rejects when destroy() is called before the iframe is ready', async () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      const readyPromise = client.ready();
-      client.destroy();
-      await expect(readyPromise).rejects.toThrow(/destroyed before ready/);
-    });
-
-    it('throws when no Window is available', () => {
-      vi.stubGlobal('window', undefined);
-      try {
-        expect(
-          () => new NosskeyIframeClient({ iframeUrl: 'https://nosskey.example/iframe' })
-        ).toThrow(/requires a Window/);
-      } finally {
-        vi.unstubAllGlobals();
-      }
-    });
-
-    it('throws when no Document is available', () => {
-      const windowWithoutDocument = {
-        addEventListener: () => undefined,
-        removeEventListener: () => undefined,
-      } as unknown as Window;
-      expect(
-        () =>
-          new NosskeyIframeClient({
-            iframeUrl: 'https://nosskey.example/iframe',
-            window: windowWithoutDocument,
-          })
-      ).toThrow(/requires a Document/);
-    });
-
-    it('throws when no container element can be determined', () => {
-      const docWithoutBody = { body: null } as unknown as Document;
-      expect(
-        () =>
-          new NosskeyIframeClient({
-            iframeUrl: 'https://nosskey.example/iframe',
-            window: harness.window,
-            document: docWithoutBody,
-          })
-      ).toThrow(/could not determine a container/);
-    });
-
-    it('makeId falls back to a non-crypto id when crypto.randomUUID is unavailable', () => {
-      (harness.window as unknown as { crypto: unknown }).crypto = {};
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      const pending = client.getPublicKey();
-      pending.catch(() => undefined);
-      const [request] = harness.iframes[0].contentWindow.postMessage.mock.calls[0];
-      expect(request.id).toMatch(/^nosskey-/);
-      client.destroy();
-    });
-
-    it('exposes the mounted iframe element via the iframe getter', () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      expect(client.iframe).toBe(harness.iframes[0] as unknown as HTMLIFrameElement);
-      client.destroy();
-    });
-
-    it('rejects requests issued after destroy()', async () => {
-      const client = new NosskeyIframeClient({
-        iframeUrl: 'https://nosskey.example/iframe',
-        window: harness.window,
-        document: harness.document,
-        container: harness.container as unknown as HTMLElement,
-      });
-      client.destroy();
-      await expect(client.getPublicKey()).rejects.toThrow(/has been destroyed/);
-    });
-  });
-});
